Double HM capacity when entries get too high
When the number of entries in the HashMap exceeds the product of the
HashMap capacity and the HashMap load factor, double the HashMap capacity
and re-distribute the entries.

For initial HashMap capacity of 16 and load factor of 0.75, this will
first occur when the number of entries becomes greater than 12 - the
capacity will double and become 32, and all of the previous entries will
be re-hashed and distributed amongst the 32 new buckets. This behavior
will continue subsequently every time the entries exceed capacity * load
factor.